implement user Auth

+++ services/user-service/src/main/java/com/kwetter/userservice/Controllers/UserController.java
@@ -4,16 +4,24 @@ import com.kwetter.userservice.Domain.Dto.ChangeUsernameDto;
 import com.kwetter.userservice.Domain.Dto.UserDto;
 import com.kwetter.userservice.Domain.Models.User;
 import com.kwetter.userservice.Domain.Service.UserService;
+import com.nimbusds.jose.shaded.json.JSONArray;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import org.apache.commons.lang3.reflect.FieldUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Map;
 
 @Api(tags = "User Controller")
 @RestController
 @CrossOrigin(origins = "*")
+@PreAuthorize("hasAuthority('SCOPE_User.All')")
 public class UserController {
 
     @Autowired
@@ -27,8 +35,9 @@ public class UserController {
 
     @ApiOperation("Create User")
     @PostMapping("/new")
-    public ResponseEntity<Object> createUser(@RequestBody UserDto dto) {
-        User user = service.addUser(dto);
+    public ResponseEntity<Object> createUser(@RequestBody UserDto dto) throws IllegalAccessException {
+        Authentication authContext = SecurityContextHolder.getContext().getAuthentication();
+        User user = service.addUser(dto, authContext);
         if (user != null) {
             return new ResponseEntity<>(user, HttpStatus.CREATED);
         } else {
@@ -39,8 +48,9 @@ public class UserController {
     @ApiOperation("Update User")
     @PostMapping("/changeUsername")
     public ResponseEntity<Object> updateUser(@RequestBody ChangeUsernameDto dto) {
+        Authentication authContext = SecurityContextHolder.getContext().getAuthentication();
         try {
-            User user = service.changeUserName(dto);
+            User user = service.changeUserName(dto, authContext);
             return new ResponseEntity<>(user, HttpStatus.CREATED);
         } catch (Exception e) {
             return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
@@ -49,10 +59,25 @@ public class UserController {
 
     @ApiOperation("Delete user")
     @DeleteMapping("/delete/{id}")
-    public HttpStatus deleteUser(@PathVariable String id) {
-        if (service.deleteUser(id)) {
-            return HttpStatus.valueOf(200);
+    public HttpStatus deleteUser(@PathVariable String id) throws IllegalAccessException {
+        Authentication authContext = SecurityContextHolder.getContext().getAuthentication();
+        if (isAdmin(authContext)) {
+            if (service.deleteUser(id)) {
+                return HttpStatus.valueOf(200);
+            }
+            return HttpStatus.BAD_REQUEST;
         }
-        return HttpStatus.BAD_REQUEST;
+        else return HttpStatus.FORBIDDEN;
+    }
+
+    private boolean isAdmin(Authentication authContext) throws IllegalAccessException {
+        Map<?, ?> claims = (Map<?, ?>) FieldUtils.readField(authContext.getPrincipal(), "claims", true);
+        JSONArray arr = (JSONArray) claims.get("roles");
+        if (arr != null) {
+            String role = (String) arr.get(0);
+            return role.equals("Role.Admin");
+        } else {
+            return false;
+        }
     }
 }
